feat(blackjack) : [분리] Winner object
